# Fixes #17921 - optimized task queries