Task queries were optimized for not bringing too much unneeded
objects. For more optimizations a proper performance analysis
is needed