@iNecas 2017-01-16
What about adding some test with benchmarking?