@ShimShtein 2017-01-17
Benchmark tests added. To run them use:
```

RAILS_ENV=production ruby -Itest ../foreman_remote_execution/test/benchmark/run_hosts_job_benchmark.rb

RAILS_ENV=production ruby -Itest ../foreman_remote_execution/test/benchmark/targeting_benchmark.rb
```
from foreman folder.