@ShimShtein 2017-01-17
## Benchmarking results on my machine:
### after changes:

```
Warming up --------------------------------------
rex-targeting-resolve-hosts
                         1.000  i/100ms
Calculating -------------------------------------
rex-targeting-resolve-hosts
                        229.277  (±14.8%) i/s -      2.197k in   9.999769s
Memory stats
Total objects allocated: 42017977
Total heap pages allocated: 6942
```

### Before changes

```
Warming up --------------------------------------
rex-targeting-resolve-hosts
                         1.000  i/100ms
Calculating -------------------------------------
rex-targeting-resolve-hosts
                        179.630  (±10.6%) i/s -      1.753k in  10.000234s
Memory stats
Total objects allocated: 42724693
Total heap pages allocated: 6929
```

## Hosts job benchmarking

### after changes:

```
Warming up --------------------------------------
   rex-run-hosts-job     1.000  i/100ms
Calculating -------------------------------------
   rex-run-hosts-job     94.729  (±12.7%) i/s -    920.000  in  10.006169s
Memory stats
Total objects allocated: 45232408
Total heap pages allocated: 6847
```

### Before change (to the hosts_job file only)

```
Warming up --------------------------------------
   rex-run-hosts-job     1.000  i/100ms
Calculating -------------------------------------
   rex-run-hosts-job     94.087  (±10.6%) i/s -    919.000  in  10.007119s
Memory stats
Total objects allocated: 45223668
Total heap pages allocated: 6847
```