@iNecas 2017-01-18
The difference doesn't seem that significant from this numbers, unless I'm reading something wrong? I guess we need;

1) verify with production setup (including Katello)
2) looking into other potential places for improvement.