@ShimShtein 2017-01-19
Basically agreed to both.

From the results, I can see that the first change (targeting) is pretty efficient - total time decrease from ~180 hosts per second to ~230 hosts per second. Plus the memory footprint went down by 700K objects.

The second change (run_hosts_job) is indeed pretty useless (it has improved memory usage by 10K objects, but it's marginally an accomplishment).

I will try to see if there is something that I am missing in run_hosts_job usecase, otherwise I will remove the unnecessary optimization.