@iNecas 2017-01-30
Works nicely and the benchmark proves improvement (although I expected a bit higher numbers). We need to have eyes opened for further non-efficiencies. Pulling this in for now. Thanks @ShimShtein